+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/index.js":"1","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/App.js":"2","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/reportWebVitals.js":"3","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/CampusCard.js":"4","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/reducers/index.js":"5","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/reducers/actionTypes.js":"6","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/store.js":"7"},{"size":627,"mtime":1611520374763,"results":"8","hashOfConfig":"9"},{"size":200,"mtime":1611529760284,"results":"10","hashOfConfig":"9"},{"size":362,"mtime":1611517037702,"results":"11","hashOfConfig":"9"},{"size":1108,"mtime":1611542158596,"results":"12","hashOfConfig":"9"},{"size":836,"mtime":1611532286891,"results":"13","hashOfConfig":"9"},{"size":51,"mtime":1611518231760,"results":"14","hashOfConfig":"9"},{"size":275,"mtime":1611517702256,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"uu5sh0",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/index.js",[],["31","32"],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/App.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/reportWebVitals.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/CampusCard.js",["33"],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/reducers/index.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/reducers/actionTypes.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/store.js",[],{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","severity":1,"message":"39","line":24,"column":43,"nodeType":"40","messageId":"41","endLine":24,"endColumn":45},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],"eqeqeq","Expected '!==' and instead saw '!='.","BinaryExpression","unexpected","no-global-assign","no-unsafe-negation"]
+[{"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/index.js":"1","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/App.js":"2","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/reportWebVitals.js":"3","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/CampusCard.js":"4","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/reducers/index.js":"5","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/reducers/actionTypes.js":"6","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/store.js":"7","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/CampusPage.js":"8","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/IsLoading.js":"9","/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/index.js":"10"},{"size":627,"mtime":1611520374763,"results":"11","hashOfConfig":"12"},{"size":200,"mtime":1611545437032,"results":"13","hashOfConfig":"12"},{"size":362,"mtime":1611517037702,"results":"14","hashOfConfig":"12"},{"size":298,"mtime":1611552642245,"results":"15","hashOfConfig":"12"},{"size":834,"mtime":1611550214548,"results":"16","hashOfConfig":"12"},{"size":51,"mtime":1611518231760,"results":"17","hashOfConfig":"12"},{"size":275,"mtime":1611517702256,"results":"18","hashOfConfig":"12"},{"size":1234,"mtime":1611552584783,"results":"19","hashOfConfig":"12"},{"size":215,"mtime":1611550880861,"results":"20","hashOfConfig":"12"},{"size":105,"mtime":1611552512766,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},"uu5sh0",{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/index.js",[],["43","44"],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/App.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/reportWebVitals.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/CampusCard.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/reducers/index.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/reducers/actionTypes.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/redux/store.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/CampusPage.js",["45"],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/IsLoading.js",[],"/Users/adinascheinfeld/Desktop/TTP/Campus-CRUD/Campus-CRUD-Frontend/crud-frontend/src/components/index.js",[],{"ruleId":"46","replacedBy":"47"},{"ruleId":"48","replacedBy":"49"},{"ruleId":"50","severity":1,"message":"51","line":16,"column":45,"nodeType":"52","messageId":"53","endLine":16,"endColumn":47},"no-native-reassign",["54"],"no-negated-in-lhs",["55"],"eqeqeq","Expected '!==' and instead saw '!='.","BinaryExpression","unexpected","no-global-assign","no-unsafe-negation"]
