+++ src/components/IsLoading.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class IsLoading extends Component {
+    render() {
+        return(
+            <div>
+                Loading
+            </div>
+        )
+    }
+}
+
+export default IsLoading;
