+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as CampusCard } from './CampusCard';
+export { default as IsLoading } from './IsLoading';
